+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Export `FiltersForm` standalone
 * New icons for Studio
 * Updating sizing for various components
+* Keyboard shortcuts for Selects
+* Remove old `Filter` component
 
 # 1.0.0